Return correct error codes from bumpfee()
The `bumpfee()` RPC was returning misleading or incorrect error codes
(for example RPC_INVALID_ADDRESS_OR_KEY when the transaction was not
BIP125 replacable. This PR changes those error codes to what I think are
more sensible:

- RPC_INVALID_PARAMETER if a single parameter is incorrect
- RPC_WALLET_ERROR for any other error, for example if the transaction
  has descendants in the mempool or the change output is too small.

This PR also updates the test cases to explicitly test the error code.