[logging] Call ShrinkDebugFile() from within OpenDebugLog()
This is mostly code move only. ShrinkDebgFile() is only called from one
place - AppInitMain(), immediately before calling OpenDebugLog(). It
makes sense to have OpenDebugLog() be responsible for calling
ShrinkDebugFile().

There is a slight difference in behaviour - ShrinkDebugFile() will now
not be called if fPrintToDebugLog is false. This is only the case for
the unit tests and bench tests, so this change is ok.