Support having SegWit always active in regtest
Most tests shouldn't have to deal with the now-historical SegWit activation
transition (and other deployments, but SegWit is certainly the hardest one
to accomodate).

This commit makes a versionbits starttime of -1 equal to "always active",
and enables it by default for SegWit on regtest. Individual tests can
override this by using the existing `-vbparams` option.

A few unit tests and functional tests are adapted to indeed override
vbparams, as they specifically test the transition.