[net processing] Don't cache local start height
Currently we cache our best block's height in CConnman in the
PeerLogicValidation::UpdatedBlockTip() callback. That cached best block
height is copied to the CNode on construction and then used when
constructing our VERSION message. The value is not used again.

Instead, just query validation for the current best height when
constructing the VERSION message.