[net processing] remove fGetAddr
fGetAddr is a flag indicating whether a GETADDR request to this peer is
outstanding. It was originally introduced in commit dd519206 where it
would be cleared on receipt of any ADDR message. Presumably the intent
was to make sure that on receipt of an ADDR response to a GETADDR, we
wouldn't blast out the received address records to our peers using our
address relay logic.

f93d5f9f changed the logic so that the flag would only get cleared if
the ADDR message contained fewer than 1000 records. It's unclear why
this change was made. The response to a GETADDR message may contain 1000
(MAX_ADDR_TO_SEND) records, in which case the flag wouldn't get cleared
until a subsequent ADDR message had been received.

5cbf7532 also changed the logic so that we'd only relay address records
if the ADDR message contained at most 10 records. This may have been
some mitigation against spam/DoS messages, and is also a good heuristic
for whether the ADDR message is a response to a GETADDR (GETADDR
responses will typically contain 100s of records).

fGetAddr has therefore outlived its usefullness. Remove it.