[net] Don't consider fRelayTxes and pfilter in eviction logic
Inbound peer eviction logic should determine a set of the 'best' peers
for us. We consider most recent block delivered, most recent tx
delivered, ping time, etc. Whether or not a peer has requested that we
deliver them transactions, or has sent us a bloom filter should not be
part of that consideration.