Avoid sending second self-announcement too soon
Mark the time of first self-announcement (before VERACK)
as if it is regular, so that the second self-announcement will
be sent out only after the time passes.

Note that since the first self-announcement is made at the end
of processing VERSION, at that point addrLocal would be already
filled with what the peers tells us we are, so this self-announcement
would be no worse than any further one.